+++ http4k-client-okhttp/src/main/kotlin/org/http4k/client/OkHttp.kt
@@ -3,13 +3,14 @@ package org.http4k.client
 import okhttp3.OkHttpClient
 import okhttp3.RequestBody.create
 import okhttp3.internal.http.HttpMethod.permitsRequestBody
+import org.http4k.core.BodyMode
 import org.http4k.core.HttpHandler
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status
 import java.net.SocketTimeoutException
 
-class OkHttp(private val client: OkHttpClient = defaultOkHttpClient(), private val bodyMode: ResponseBodyMode = ResponseBodyMode.Memory) : HttpHandler {
+class OkHttp(private val client: OkHttpClient = defaultOkHttpClient(), private val bodyMode: BodyMode.Response = BodyMode.Response.Memory) : HttpHandler {
 
     private fun Request.asOkHttp(): okhttp3.Request =
         headers.fold(okhttp3.Request.Builder()