+++ http4k-client-okhttp/src/test/kotlin/org/http4k/client/OkHttpStreamingTest.kt
@@ -1,6 +1,7 @@
 package org.http4k.client
 
+import org.http4k.core.BodyMode
 import org.http4k.server.SunHttp
 
-class OkHttpStreamingTest : Http4kClientContract({ SunHttp(it) }, OkHttp(bodyMode = ResponseBodyMode.Stream),
-    OkHttp(timeout, bodyMode = ResponseBodyMode.Stream))
+class OkHttpStreamingTest : Http4kClientContract({ SunHttp(it) }, OkHttp(bodyMode = BodyMode.Response.Stream),
+    OkHttp(timeout, bodyMode = BodyMode.Response.Stream))