+++ http4k-core/src/test/kotlin/org/http4k/streaming/SunHttpStreamingTest.kt
@@ -1,7 +1,7 @@
 package org.http4k.streaming
 
 import org.http4k.client.ApacheClient
-import org.http4k.client.ResponseBodyMode
+import org.http4k.core.BodyMode
 import org.http4k.core.HttpHandler
 import org.http4k.server.ServerConfig
 import org.http4k.server.SunHttp
@@ -9,6 +9,7 @@ import org.http4k.server.SunHttp
 class SunHttpStreamingTest : StreamingContract() {
     override fun serverConfig(port: Int): ServerConfig = SunHttp(port)
 
-    override fun createClient(): HttpHandler = ApacheClient(bodyMode = ResponseBodyMode.Stream)
+    override fun createClient(): HttpHandler =
+        ApacheClient(requestBodyMode = BodyMode.Request.Stream, responseBodyMode = BodyMode.Response.Stream)
 
 }