+++ http4k-server-undertow/src/test/kotlin/org/http4k/stream/UndertowStreamingTest.kt
@@ -1,7 +1,7 @@
 package org.http4k.stream
 
 import org.http4k.client.ApacheClient
-import org.http4k.client.ResponseBodyMode
+import org.http4k.core.BodyMode
 import org.http4k.core.HttpHandler
 import org.http4k.server.ServerConfig
 import org.http4k.server.Undertow
@@ -10,6 +10,7 @@ import org.http4k.streaming.StreamingContract
 class UndertowStreamingTest : StreamingContract() {
     override fun serverConfig(port: Int): ServerConfig = Undertow(port)
 
-    override fun createClient(): HttpHandler = ApacheClient(bodyMode = ResponseBodyMode.Stream)
+    override fun createClient(): HttpHandler =
+        ApacheClient(requestBodyMode = BodyMode.Request.Stream, responseBodyMode = BodyMode.Response.Stream)
 
 }